requirements.txt
pytest==7.4.0
selene==2.0.0rc9
allure-pytest==2.13.2
selenium==4.31.0
allure-python-commons~=2.13.2
dotenv